hide jwt-go functionality behind new jwt package interface.
